Break out AutoEquipPatches
